push sang

+++ views/employees/users/list.php
@@ -24,38 +24,49 @@
                     <th>Thao tác</th>
                   </thead>
                   <tbody>
-                    <?php foreach($items as $key => $item): ?>
-                        <tr>
-                          <?php
-                              $username = $item['username'];
-                              $fullname = $item['fullname'];
-                              $deptname = $item['deptname'];
-                              $status   = ($item['enable'] == 'active') ? 'Kích hoạt' : 'Không kích hoạt';
-                          ?>
-                            <td><?php echo $key + 1  ?></td>
-                            <td><?php echo $fullname  ?></td>
-                            <td><?php echo $username  ?></td>
-                            <td><?php echo $deptname  ?></td>
-                            <td><?php echo $status  ?></td>
-                            <td>
-                              <?php 
-                                  $editLink = "?controller=" . $this->controllerName . "&action=update&id=" .$item['id'];
-                                  $editChangeStatus = "?controller=" . $this->controllerName . "&action=changeStatus&id=" .$item['id'] . "&enable=" . $item['enable'];
-                                  $editDelete = "?controller=" . $this->controllerName . "&action=delete&id=" .$item['id'];
-
-                              ?>
-                             <a href="<?php echo $editLink ?>" class="btn btn-primary">
-                                <i class="fa fa-pencil-alt"></i>
-                              </a>
-                              <a href="<?php echo $editChangeStatus ?>" class="btn btn-danger">
-                                <i class="fa fa-sync"></i>
-                              </a>
-                              <a href="<?php echo $editDelete ?>" class="btn btn-success">
-                                <i class="fa fa-trash"></i>
-                              </a>
-                            </td>
-                        </tr>
-                    <?php endforeach; ?>
+                    <?php if(count($items) > 0): ?>
+                      <?php foreach($items as $key => $item): ?>
+                          <tr>
+                            <?php
+                                $username = $item['username'];
+                                $fullname = $item['fullname'];
+                                $deptname = $item['deptname'];
+                                $status   = ($item['enable'] == 'active') ? 'Kích hoạt' : 'Không kích hoạt';
+                            ?>
+                              <td><?php echo $key + 1  ?></td>
+                              <td><?php echo $fullname  ?></td>
+                              <td><?php echo $username  ?></td>
+                              <td><?php echo $deptname  ?></td>
+                              <td><?php echo $status  ?></td>
+                              <td>
+                                <?php 
+                                    $editLink = "?controller=" . $this->controllerName . "&action=update&id=" .$item['id'];
+                                    $editChangeStatus = "?controller=" . $this->controllerName . "&action=changeStatus&id=" .$item['id'] . "&enable=" . $item['enable'];
+                                    $editDelete = "?controller=" . $this->controllerName . "&action=delete&id=" .$item['id'];
+                                    $resetPasswordLink = "?controller=" . $this->controllerName . "&action=resetPassword&id=" .$item['id'];
+                                ?>
+                                <a href="<?php echo $resetPasswordLink ?>" class="btn btn-default">
+                                <i class="fa fa-cog"></i>
+                                </a>
+                                <a href="<?php echo $editLink ?>" class="btn btn-primary">
+                                  <i class="fa fa-pencil-alt"></i>
+                                </a>
+                                <a href="<?php echo $editChangeStatus ?>" class="btn btn-danger">
+                                  <i class="fa fa-sync"></i>
+                                </a>
+                                <a href="<?php echo $editDelete ?>" class="btn btn-success">
+                                  <i class="fa fa-trash"></i>
+                                </a>
+                              </td>
+                          </tr>
+                      <?php endforeach; ?>
+                    <?php else: ?>
+                      <tr>
+                        <td colspan="6">
+                          <b>Không tồn tại bản ghi </b>
+                        </td>
+                      </tr>
+                    <?php endif ?>
                   </tbody>
                 </table>
               </div>
